Преобразование типов, чтобы избавиться от предупреждений на C++ Builder 2010

+++ arch/cbuilder/MxChart.cpp
@@ -5199,8 +5199,10 @@ void irs::chart::builder_chart_window_t::controls_t::update_chart()
 
       TMxChartItem* item = mp_chart->Items[mp_chart->BaseItem];
       if (!m_fixed_time_mode_enabled) {
-        const double first = item->DataX[item->Bounds.Begin];
-        const double last = item->DataX[item->Bounds.End];
+        const double first = static_cast<double>(item->DataX[
+          static_cast<int>(item->Bounds.Begin)]);
+        const double last = static_cast<double>(item->DataX[
+          static_cast<int>(item->Bounds.End)]);
         area.Left = last - (area.Right - area.Left);
         area.Right = last;
       }
